requirements.txt
Django==1.11.15
django-filter==2.0.0
djangorestframework==3.8.2
